Add `prefix` argument to `supercell` docstring
The `supercell` function already takes a `prefix` argument, but I realized that it is missing from the docstring. This means that from within PyMOL, users don't realize that this parameter can be changed!

The ability to change prefixes is especially important because otherwise, running `supercell` a second time in the same PyMOL session will overwrite the previous symmetry mates `m001_1`, `m001_2`, etc.